requirements.txt
fastapi~=0.115.13
rembg>=2.0.0
pillow>=9.0.0
python-multipart>=0.0.5
uvicorn>=0.15.0
onnxruntime>=1.15.0
jinja2>=3.0.0
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
